+++ bench/random.ts
